refactor: move manager image upload core to service layer
